+++ plugins/procedural/main.cpp
@@ -488,7 +488,15 @@ scene_load
         AtString renderSettings;
         if (AiParamValueMapGetStr(params, str::render_settings, &renderSettings) && renderSettings.length() > 0)
             reader->SetRenderSettings(std::string(renderSettings.c_str()));
-        
+        // We use the "render_pass" argument to set the path to the render pass to use for rendering
+        // With kick you can use -sceneload_arg to pass the argument, note that is works only with the scene index enabled (hydra2)
+        //
+        // kick -sceneload_arg render_pass /Render/Passes/renderpass2 myscene.usda
+        //
+        AtString renderPass;
+        if (AiParamValueMapGetStr(params, str::render_pass, &renderPass) && renderPass.length() > 0) {
+            reader->SetRenderPass(std::string(renderPass.c_str()));
+        }
     }
     reader->SetFrame(frame);